SCHED-535: Moving logic.
