+++ scheduler/core/builder/__init__.py
@@ -1,5 +1,7 @@
 # Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-from .builder import *
 from .blueprint import *
+from .operationsbuilder import *
+from .schedulerbuilder import *
+from .validationbuilder import *
