+++ scheduler/core/builder/operationsbuilder.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2016-2023 Association of Universities for Research in Astronomy, Inc. (AURA)
+# For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+from .schedulerbuilder import SchedulerBuilder
+
+
+class OperationBuilder(SchedulerBuilder):
+    ...
