--- tests/unit/scheduler/core/collector/collector_fixture.py
@@ -1,38 +1,0 @@
-
-import pytest
-
-from astropy.time import Time
-from lucupy.minimodel import ALL_SITES, NightIndex, Semester, SemesterHalf
-
-from scheduler.core.builder.blueprint import CollectorBlueprint
-from scheduler.core.builder.builder import ValidationBuilder
-from scheduler.core.sources import Sources
-from scheduler.core.eventsqueue import EventQueue
-
-
-@pytest.fixture
-def scheduler_collector():
-    start = Time("2018-10-01 08:00:00", format='iso', scale='utc')
-    end = Time("2018-10-03 08:00:00", format='iso', scale='utc')
-    num_nights_to_schedule = 3
-    sites = ALL_SITES
-    semesters = frozenset([Semester(2018, SemesterHalf.B)])
-
-    collector_blueprint = CollectorBlueprint(
-        obs_class=['SCIENCE', 'PROGCAL', 'PARTNERCAL'],
-        prg_type=['Q', 'LP', 'FT', 'DD'],
-        time_slot_length=1.0
-    )
-
-    night_indices = frozenset(NightIndex(idx) for idx in range(num_nights_to_schedule))
-    builder = ValidationBuilder(Sources(), EventQueue(night_indices, sites))
-
-    collector = builder.build_collector(
-        start=start,
-        end=end,
-        sites=sites,
-        semesters=semesters,
-        blueprint=collector_blueprint
-    )
-
-    return collector
